% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{corrExpPredPlot}
\alias{corrExpPredPlot}
\title{Generate correlation plot between predicted and expected cell type
proportions.}
\usage{
corrExpPredPlot(
  object,
  facet.by,
  color.by,
  colors,
  corr = "pearson",
  filter.sc = TRUE,
  pos.x.label = 0.01,
  pos.y.label = 0.95,
  ncol = NULL,
  nrow = NULL,
  title = NULL,
  theme = theme_grey()
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{eval.stats.samples} slot.}

\item{facet.by}{Variable used for showing data in different panels. If it is
\code{NULL}, the plot is not separated into different panels. The options
are \code{nMix} (by number of different cell types) and \code{CellType}
(by cell type).}

\item{color.by}{Variable used to color data. The options
are \code{nMix} (by number of different cell types) and \code{CellType}
(by cell type).}

\item{colors}{Vector with colors for the plot.}

\item{ncol}{Number of columns if \code{facet.by} is different than \code{NULL}.}

\item{nrow}{Number of rows if \code{facet.by} is different than \code{NULL}.}

\item{title}{Title of the plot.}

\item{ylimit}{Upper limit in y axis if it is needed. \code{NULL} by default.}
}
\description{
Generate correlation plot between predicted and expected cell type
proportions.
}
\examples{
DDLSChung <- violinError(
  object = DDLSChung,

)

}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
