% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvDigitalDLSorter}
\alias{deconvDigitalDLSorter}
\title{Deconvolute bulk gene expression samples (RNA-Seq) using a pre-trained
DigitalDLSorter model.}
\usage{
deconvDigitalDLSorter(
  data,
  model = "breast",
  batch.size = 128,
  transpose = FALSE,
  normalize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A matrix bulk gene expression with}

\item{batch.size}{Number of samples per gradient update. If unspecified,
\code{batch.size} will default to 128.}

\item{verbose}{Boolean indicating if show the progression of the model during
training. Beisdes, it is shown information about the architecture of the model
(\code{TRUE} by default).}

\item{numm.epochs}{Number of epochs to train the model.}

\item{val}{Boolean that determine if a validation subset is used during
training (\code{FALSE} by default).}

\item{freq.val}{Number between 0.1 and 0.5 that determine the number of
samples from training data that will be used as validation subset.}

\item{loss}{Character indicating loss function selected for training the model
(Kullback-Leibler divergence by default).
See \code{keras} documentation for more details.}

\item{metrics}{Vector fo metrics used to evaluate the performance of the model
during training and on test data (\code{c("accuracy", "mean_absolute_error",
"categorical_accuracy")} by default)}

\item{view.metrics.plots}{Boolean indicating if show progression plots of loss
and metrics during training (\code{TRUE} by default). \code{keras} for R allows
to see the progression of the model during training if you are working in RStudio.}
}
\value{
A DigitalDLSorter object with \code{trained.model} slot containing
a \code{\link{DigitalDLSorterDNN}} object. For more information about the
structure of this class, see \code{\link{?DigitalDLSorterDNN}}.
}
\description{
Deconvolute bulk gene expression samples (RNA-Seq) enumerating and
quantifying the proportion of cell types present in a bulk sample.
}
\details{
This method uses a pre-trained Deep Neural Network model to enumerate
and quantify the cell types present in bulk RNA-Seq samples. For the moment,
the available models allow to deconvolute the immune infiltration in colorectal
(\code{colon.li} (Li et al., 2017)) and breast cancer (\code{breast.chung}
(Chung et al., 2017)).
}
\examples{
DDLSChung <- trainDigitalDLSorterModel(
  object = DDLSChung,
  batch.size = 128,
  num.epochs = 20
)

}
\references{
Chung, W., Eum, H. H., Lee, H. O., Lee, K. M., Lee, H. B., Kim,
K. T., et al. (2017). Single-cell RNA-seq enables comprehensive tumour and
immune cell profiling in primary breast cancer. Nat. Commun. 8 (1), 15081.
doi: \url{10.1038/ncomms15081}.
}
\seealso{
\code{\link{plotTrainingHistory}}, \code{\link{deconvDigitalDLSorterModel}}.
}
