% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvDigitalDLSorterObj}
\alias{deconvDigitalDLSorterObj}
\title{Deconvolute bulk gene expression samples (RNA-Seq).}
\usage{
deconvDigitalDLSorterObj(
  object,
  name.data,
  batch.size = 128,
  transpose = FALSE,
  normalize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\link{DigitalDLSorter}} object with \code{trained.data}
and \code{deconv.data} slots.}

\item{name.data}{Name of the data store in \code{DigitalDLSorter} object. If
it is not provided, the first dataset will be used.}

\item{batch.size}{Number of samples per gradient update. If unspecified,
\code{batch.size} will default to 128.}

\item{transpose}{Transpose data for set features (genes) as columns and
samples as rows. \code{TRUE} by default.}

\item{normalize}{Normalize data before deconvolution. \code{TRUE} by default.}

\item{verbose}{Show messages during the execution.}
}
\value{
A \code{data.frame} with samples (\eqn{i}) as rows and cell types
(\eqn{j}) as columns. Each entry represents the proportion of \eqn{j} cell
type in \eqn{i} sample.
}
\description{
Deconvolute bulk gene expression samples (RNA-Seq) enumerating and
quantifying the proportion of cell types present in a bulk sample. This
function needs a \code{DigitalDLSorter} object with a trained DNN model
(\code{\link{trained.model}} slot) and data for deconvoluting store in
\code{deconv.data} slot.
}
\details{
This function is oriented for users that have trained a DNN model using their
own data. If you want to use a pre-trained model, see
\code{\link{deconvDigitalDLSorter}}.
}
\examples{
DDLSChung <- trainDigitalDLSorterModel(
  object = DDLSChung,
  batch.size = 128,
  num.epochs = 20
)

}
\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify
immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi:
\url{10.3389/fgene.2019.00978}
}
\seealso{
\code{\link{trainDigitalDLSorter}} \code{\link{DigitalDLSorter}}
}
