% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{distErrorPlot}
\alias{distErrorPlot}
\title{Generate boxplot or violin plot showing how errors are distributed.}
\usage{
distErrorPlot(
  object,
  error,
  colors,
  x.by = "pBin",
  facet.by = "nMix",
  color.by = "nMix",
  filter.sc = TRUE,
  error.labels = FALSE,
  pos.x.label = 4.6,
  pos.y.label = NULL,
  size.point = 0.1,
  alpha.point = 1,
  type = "violinplot",
  ylimit = NULL,
  nrow = NULL,
  ncol = NULL,
  title = NULL,
  theme = theme_grey(),
  ...
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot
containing metrics in \code{eval.stats.samples} slot.}

\item{error}{Which error is going to represent. The available errors are
absolute error (\code{"AbsErr"}), proportional absolute error
(\code{"ppAbsErr"}), squared error (\code{"SqrErr"}) or proportional
squared error (\code{"ppSqrErr"}).}

\item{colors}{Vector of colors to use. Only vectors with a number of colors
equal to or greater than the levels of \code{color.by} will be accepted.
By default is used a list of custom colors provided by us.}

\item{x.by}{Variable used for x axis. When \code{facet.by} is not \code{NULL},
the best option is \code{pBin} (probability bin). The options
are \code{nMix} (by number of different cell types), \code{CellType}
(by cell type) and \code{pBin}.}

\item{facet.by}{Variable used for showing data in different panels. If it is
\code{NULL}, the plot is not separated into different panels. The options
are \code{nMix} (by number of different cell types) and \code{CellType}
(by cell type).}

\item{color.by}{Variable used to color data. The options
are \code{nMix} (by number of different cell types) and \code{CellType}
(by cell type).}

\item{filter.sc}{Filter prediction errors of single-cell profiles. If
\code{TRUE}, (by default) only will be shown predition errors of bulk samples.}

\item{ylimit}{Upper limit in y axis if it is needed. \code{NULL} by default.}

\item{nrow}{Number of rows if \code{facet.by} is different than \code{NULL}.}

\item{ncol}{Number of columns if \code{facet.by} is different than \code{NULL}.}

\item{title}{Title of the plot.}

\item{theme}{Ggplot theme.}
}
\description{
Generate violin plot showing how errors are distributed. There are different
possibilities of showing the results. See examples for more information.
}
\examples{
DDLSChung <- violinError(
  object = DDLSChung,

)

}
\seealso{
\code{\link{calculateEvalMetrics}} \code{\link{corrExpPredPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
