% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSingleCell.R
\name{estimateZinbwaveParams}
\alias{estimateZinbwaveParams}
\title{Estimate parameteres for ZINB-WaVE model for simulating single-cell expression
profiles.}
\usage{
estimateZinbwaveParams(
  object,
  cell.ID.column,
  gene.ID.column,
  cell.type.column,
  cell.cov.columns,
  gene.cov.columns,
  set.type = "All",
  threads = 2,
  verbose = TRUE
)
}
\arguments{
\item{object}{DigitalDLSorter object with a \code{single.cell.real} slot.}

\item{cell.ID.column}{Name or number of the column in cells.metadata
corresponding with cell names in expression matrix.}

\item{gene.ID.column}{Name or number of the column in genes.metadata
corresponding with the nonation used for features/genes.}

\item{cell.type.column}{Name or number of the column in cells.metadata
corresponding with the cell type of each cell.}

\item{cell.cov.columns}{Name or number of columns in cells.metadata that
will be used as covariates in the model.}

\item{gene.cov.columns}{Name or number of columns in genes.metadata that will
be used as covariates in the model.}

\item{set.type}{Cell type to evaluate. 'All' by default.}

\item{threads}{Number of threads used for the estimation. For setting this environment
\code{BiocParallel} package is used.}

\item{verbose}{Show messages during the execution}
}
\value{
A DigitalDLSorter object with \code{zinb.params} slot with a ZinbParams
object containing the estimated ZINB-WaVE parameters from real single-cell data.
}
\description{
Estimate simulation parameters for the ZINB-WaVE model from a real single-cell
dataset.
}
\details{
ZINB-WaVE is a flexible model for zero-inflated count data. This function
carries out the model fit to real single-cell data modeling \eqn{Y_{ij}}
(the count of feature \eqn{j} for sample \eqn{i}) as a random variable
following a zero-inflated negative binomial (ZINB) distribution. The estimated
paramters will be used for the simulation of new single-cell expression profiles.
To do this, we use \code{zinbEstimate} function from \code{splatter} package,
that is a wrapper around \code{zinbFit} function from \code{zinbwave} package.
For details about the model, see Risso et al. 2018.
}
\examples{
DDLSChung <- estimateZinbwaveParams(
  object = DDLSChung,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "external_gene_name",
  cell.type.column = "Cell_type",
  cell.cov.columns = c("Patient", "Sample_type"),
  gene.cov.columns = "gene_length",
  threads = 4,
  verbose = TRUE
)

}
\seealso{
\code{\link{simSingleCellProfiles}}
}
