% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{eval.stats}
\alias{eval.stats}
\alias{eval.stats<-}
\title{Get and set \code{eval.stats} slot in a \code{DigitalDLSorterDNN}
object.}
\usage{
eval.stats(object)

eval.stats(object) <- value
}
\arguments{
\item{object}{A \code{DigitalDLSorterDNN} object.}

\item{value}{A \code{list} object with the resulting metrics after prediction
on test data with DNN model.}
}
\description{
Get and set \code{eval.stats} slot in a \code{DigitalDLSorterDNN}
object.
}
