% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{loadDeconvDataFromFile}
\alias{loadDeconvDataFromFile}
\title{Load data to deconvolute from tabulated text file.}
\usage{
loadDeconvDataFromFile(object, file.path, name.data = NULL)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot.}

\item{file.path}{File path where data is stored.}

\item{name.data}{Name with which the data is stored in \code{DigitalDLSorter}
object. If \code{name.data} is not provided, basename of file is used.}
}
\description{
Load data to deconvolute from text file. Accepted formats are tsv and tas.gz.
}
\examples{

}
\seealso{
\code{\link{trainDigitalDLSorterModel}}, \code{\link{deconvDigitalDLSorterModel}}
}
