% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadRealSCProfiles}
\alias{loadRealSCProfiles}
\title{Load real scRNA-Seq data into a \code{DigitalDLSorter} object for simulating
new profiles.}
\usage{
loadRealSCProfiles(
  single.cell.real,
  cell.ID.column = 1,
  gene.ID.column = 1,
  min.counts = 0,
  min.cells = 0,
  project = "DigitalDLSorterProject"
)
}
\arguments{
\item{single.cell.real}{If data is provided from files, \code{single.cell.real}
must be a vector with three elements: single-cell counts, cells metadata and
genes metadata. If data is provided from a \code{SingleCellExperiment} object,
single-cell counts must be in \code{assay} slot, cells metadata in
\code{colData} and genes metadata in \code{rowData}.}

\item{cell.ID.column}{Name or number of the column in cells metadata
corresponding with cell names in expression matrix.}

\item{gene.ID.column}{Name or number of the column in genes metadata
corresponding with the names used for features/genes.}

\item{min.counts}{Minimum gene counts to filter (0 by default).}

\item{min.cells}{Minimum of cells with more than min.counts (0 by default).}

\item{project}{Name of the project for \code{DigitaDLSorter} object.}
}
\description{
Load scRNA-Seq data into a \code{DigitalDLSorter} from file stored on disk
or from a \code{SingleCellExperiment} object. Provided data must be composed
by three pieces of information:
\itemize{
  \item Single-cell counts: genes in rows and cells in columns.
  \item Cells metadata: with annotations (columns) for each cell (rows).
  \item Genes metadata with annotations (columns) for each gene (rows).
}
In the case that data is provided from files, \code{single.cell.real}
argument must be a vector of three elements ordered so that the first file
corresponds to counts, the second to cells metadata and the last to genes
metadata.
On the other hand, if data is provided as \code{SingleCellExperiment}, the
object must contains single-cell counts in \code{assay} slot, cells metadata
in \code{colData} slot and genes metadata in \code{rowData}.
}
\examples{
DDLSChung <- loadRealSCProfiles(
  single.cell.real = breast.chung.data,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "external_gene_name",
  min.cells = 0,
  min.counts = 0,
  project = "Chung_etal_2017"
)

}
\seealso{
\code{\link{simSingleCellProfiles}}
}
