% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{loadTrainedModelFromH5}
\alias{loadTrainedModelFromH5}
\title{Load from HDF5 file a trained DigitalDLSorter DNN model.}
\usage{
loadTrainedModelFromH5(object, file.path, reset.slot = TRUE)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot.}

\item{file.path}{Valid file path where storing model.}

\item{reset.slot}{Remove \code{trained.slot} if it already exists.
A new DigitalDLSorterDNN object will be formed, but it will not contain other
slots.}
}
\description{
Load from HDF5 file a trained DigitalDLSorter DNN model and incorporate
it in DigitalDLSorter object.
}
\examples{

}
\seealso{
\code{\link{trainDigitalDLSorterModel}}, \code{\link{deconvDigitalDLSorterModel}},
\code{\link{loadTrainedModel}}
}
