% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{plotTrainingHistory}
\alias{plotTrainingHistory}
\title{Plot training history of a trained DigitalDLSorter DNN model.}
\usage{
plotTrainingHistory(
  object,
  title = "History of metrics during training",
  metrics = NULL
)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot.}

\item{title}{Title of plot.}

\item{metrics}{Which metrics to plot. If it is equal to \code{NULL} (by default),
all metrics will be plotted.}
}
\description{
Plot training history of a trained DigitalDLSorter DNN model.
}
\examples{

}
\seealso{
\code{\link{trainDigitalDLSorterModel}}, \code{\link{deconvDigitalDLSorterModel}}
}
