% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes-class}
\alias{ProbMatrixCellTypes}
\title{The Class ProbMatrixCellTypes.}
\description{
The ProbMatrixCellTypes class is a data storage class that stores the probability
matrix used for the simulation of pseudo-bulk samples. This matrix corresponds
with \code{prob.matrix} slot. The rest of slots are additional
information generated during the process.
}
\details{
As described in Torroja and Sanchez-Cabo, 2019,
the probabilities are built by six different methods in order to avoid biases
due to the composition of the bulk samples. In \code{plots} slot are stored
a representation of these probabilities with the aim of offering a method for
monitoring the different sets generated during the process. These plots can
be displayed with \code{\link{showProbPlot}} function. See documentation for details.
}
\section{Slots}{

\describe{
\item{\code{prob.matrix}}{Matrix of probabilities generated for the simulation of
bulk samples. Rows correspond with bulk samples which will be generated (\eqn{i}),
columns are the cell types present on single-cell data provided (\eqn{j}) and each entry
is the proportion of \eqn{i} cell type on \eqn{j} sample.}

\item{\code{cell.names}}{List with the evolution of the selected metrics during
training.}

\item{\code{set.list}}{Results of the selected metrics on test data.}

\item{\code{set}}{Matrix with the deconvolution results on test data.
Columns are cell types, rows are samples and each entry is the proportion of
this cell type on this sample.}

\item{\code{exclusive.types}}{Optional slot that contains the exclusive cell types
on the experiment if they are provided. NULL by default.}

\item{\code{plots}}{List of lists with the resulting plots generated during the
construction of probabilities.}

\item{\code{type.data}}{Character with the type of data contained: training or test.}
}}

