% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{saveTrainedModel}
\alias{saveTrainedModel}
\title{Save on disk trained DigitalDLSorter DNN model.}
\usage{
saveTrainedModel(object, file.path, overwrite = FALSE)
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{trained.model} slot.}

\item{file.path}{Valid file path where storing model.}

\item{overwrite}{Overwrite file if it already exists.}
}
\description{
Save on disk the trained model in HDF5 format.
}
\examples{

}
\seealso{
\code{\link{trainDigitalDLSorterModel}}, \code{\link{loadTrainedModel}}.
}
