% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showProbPlot}
\alias{showProbPlot}
\title{Show probability plots generated with \code{\link{generateTrainAndTestBulkProbMatrix}}.}
\usage{
showProbPlot(object, type.data, set, type.plot = "maxprob")
}
\arguments{
\item{object}{\code{DigitalDLSorter} object with \code{prob.cell.types} slot
with \code{plot} slot.}

\item{type.data}{Subset of data to show: \code{train} or \code{test}.}

\item{set}{Integer determining which of the 6 different approaches used for
generating the probability matrix show.}

\item{type.plot}{Character determining which type of visualitation show.
It can be \code{boxplot}, \code{violinplot}, \code{linesplot} or
\code{maxprob}.}
}
\value{
\code{ggplot} object.
}
\description{
Show probability plots generated with \code{\link{generateTrainAndTestBulkProbMatrix}}
by different methods that will determine the proportion of different cell types
used during the simulation of bulk RNA-Seq samples. There are 6 subsets
of probabilities generated by different approaches that can be shown with three
types of visualitations: boxplot, violinplot and linesplot.
}
\details{
These plots are only for diagnostic purposes. This is the reason because they
are generated without any parameter introduced by the user.
}
\examples{
DDLSChung <- trainDigitalDLSorterModel(
  object = DDLSChung,
  batch.size = 128,
  num.epochs = 20
)

}
\seealso{
\code{\link{generateTrainAndTestBulkProbMatrix}}
}
