% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.breast.small}
\alias{TCGA.breast.small}
\title{Breast cancer bulk RNA-Seq samples from TCGA Research Network.}
\format{An object of class \code{matrix} with 44831 rows and 15 columns.}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
TCGA.breast.small
}
\description{
Subset of Breast cancer bulk RNA-Seq samples from TCGA Research Network.
FPKMs were transformed in TPMs and aggregated based on SYMBOL genes.
}
\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify
immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi:
\url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
