% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.colon}
\alias{TCGA.colon}
\title{Colorectal Bulk RNA-Seq samples from TCGA Research Network.}
\format{An object of class \code{matrix} with 59085 rows and 521 columns.}
\source{
\url{https://www.cancer.gov/tcga}
}
\usage{
TCGA.colon
}
\description{
Colorectal Bulk RNA-Seq samples from TCGA Research Network.
}
\references{
Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify
immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi:
\url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
