% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvDigitalDLSorter}
\alias{deconvDigitalDLSorter}
\title{Deconvolute bulk gene expression samples (RNA-Seq) using a pre-trained
DigitalDLSorter model.}
\usage{
deconvDigitalDLSorter(
  data,
  model = "breast",
  batch.size = 128,
  transpose = FALSE,
  normalize = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A matrix bulk gene expression with}

\item{model}{Pre-trained DNN model to use for deconvoluting process.
For the moment, the available
models are for deconvoluting RNA-Seq samples from breast cancer ('breast') and
colon cancer ('colon') environments.}

\item{batch.size}{Number of samples per gradient update. If unspecified,
\code{batch.size} will default to 128.}

\item{transpose}{Transpose data for set features (genes) as columns and
samples as rows. \code{TRUE} by default.}

\item{normalize}{Normalize data before deconvolution. \code{TRUE} by default.}

\item{verbose}{Show messages during the execution.}
}
\value{
A \code{data.frame} with samples (\eqn{i}) as rows and cell types
(\eqn{j}) as columns. Each entry represents the proportion of \eqn{j} cell
type in \eqn{i} sample.
}
\description{
Deconvolute bulk gene expression samples (RNA-Seq) enumerating and
quantifying the proportion of cell types present in a bulk sample. See in
Details the available models.
}
\details{
This method uses a pre-trained Deep Neural Network model to enumerate
and quantify the cell types present in bulk RNA-Seq samples. For the moment,
the available models allow to deconvolute the immune infiltration in colorectal
(\code{'colon.li'} (Li et al., 2017)) and breast cancer (\code{'breast.chung'}
(Chung et al., 2017)).

This function is oriented for users that only want to use the method for
deconvoluting their bulk RNA-Seq samples. For users that are buliding their
own model using the rest of functionalities of \code{digitalDLSoteR} package,
see \code{\link{deconvDigitalDLSorterObj}}. The former works with base classes,
while the last uses \code{DigitalDLSorter} objects.
}
\examples{
DDLSChung <- trainDigitalDLSorterModel(
  object = DDLSChung,
  batch.size = 128,
  num.epochs = 20
)

}
\references{
Chung, W., Eum, H. H., Lee, H. O., Lee, K. M., Lee, H. B., Kim,
K. T., et al. (2017). Single-cell RNA-seq enables comprehensive tumour and
immune cell profiling in primary breast cancer. Nat. Commun. 8 (1), 15081.
doi: \url{10.1038/ncomms15081}.
}
\seealso{
\code{\link{deconvDigitalDLSorterObj}}.
}
