% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DigitalDLSorterDNN-class}
\alias{DigitalDLSorterDNN-class}
\alias{DigitalDLSorterDNN}
\title{The DigitalDLSorterDNN Class.}
\description{
The DigitalDLSorterDNN object stores the trained Deep Neural Network, the
training history of selected metrics and the results of prediction on test data.
}
\details{
The steps realted with Depp Learning are carried out with \code{keras}
package, so the model are stored in a R6 class, system used by the package.
}
\section{Slots}{

\describe{
\item{\code{model}}{Model resulting from the training of Deep Neural Network. This slot
contains a R6 \code{keras.engine.sequential.Sequential} object with the}

\item{\code{training.history}}{List with the evolution of the selected metrics during
training.}

\item{\code{eval.stats.model}}{Results of the selected metrics on test data.}

\item{\code{predict.results}}{Matrix with the deconvolution results on test data.
Columns are cell types, rows are samples and each entry is the proportion of
this cell type on this sample.}

\item{\code{cell.types}}{Vector with the cell types to deconvolve.}

\item{\code{features}}{Vector with features used during training. These features will be
used for the folloring predictions.}
}}

