% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc.chung.breast}
\alias{sc.chung.breast}
\title{Subset from the original dataset used for generating \code{breast.chung}
model.}
\format{An object of class \code{SingleCellExperiment} with 23555 rows and 488 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75688}
}
\usage{
sc.chung.breast
}
\description{
'Toy' subset from the original dataset used for generating \code{breast.chung}
model in order to show some examples in vignette and documentation. Data
are provided as a \code{SingleCellExperiment} object with counts in
\code{assay} slot, cells metadata in \code{colData} slot and genes metadata
in \code{rowData} slot.
}
\details{
For more information about the complete dataset, see \code{?breast.chung}.
}
\references{
Chung, W., Eum, H. H., Lee, H. O., Lee, K. M., Lee, H. B., Kim,
K. T., et al. (2017). Single-cell RNA-seq enables comprehensive tumour and
immune cell profiling in primary breast cancer. Nat. Commun. 8 (1), 15081.
doi: \url{10.1038/ncomms15081}.

Torroja, C. y Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify
immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi:
\url{10.3389/fgene.2019.00978}
}
\keyword{datasets}
